Add real A100 GPU results; fix f32/PRNG tolerances + numpy-safe result saving
- Recorded real A100 80GB benchmarks for 9 tasks (9/9 correctness-gated), saved to
  results/gpu/; README now reports measured GPU + CPU numbers (two device classes).
- f32/c64 tolerance 2e-4 -> 2e-3 (matches f32 SVD/QR accuracy that scales with N);
  PRNG gets a statistical family tolerance (0.05) instead of a numeric residual tol.
- runner.save() is numpy-safe (np.float64 residuals broke json.dump).

+++ jaxbench/reference.py
@@ -76,6 +76,6 @@ def residual(op: str, inp: dict, out) -> float:
     if op in ("threefry_uniform", "threefry_normal"):
         x = np.asarray(out).ravel()
         if op == "threefry_uniform":
-            return max(abs(x.mean() - 0.5), abs(x.var() - 1/12)) * 3  # loose statistical
-        return max(abs(x.mean()), abs(x.var() - 1.0)) * 0.5
+            return float(max(abs(x.mean() - 0.5), abs(x.var() - 1/12)) * 3)  # statistical
+        return float(max(abs(x.mean()), abs(x.var() - 1.0)) * 0.5)
     raise KeyError(f"no reference for op {op!r}")
